+++ UnityAds/src/main/java/com/mopub/mobileads/UnityRouter.java
@@ -11,7 +11,6 @@ import com.mopub.common.privacy.PersonalInfoManager;
 import com.unity3d.ads.UnityAds;
 import com.unity3d.ads.metadata.MediationMetaData;
 import com.unity3d.ads.metadata.MetaData;
-import com.unity3d.services.banners.UnityBanners;
 
 import java.util.Map;
 
@@ -24,7 +23,6 @@ public class UnityRouter {
     private static final String ADAPTER_NAME = UnityRouter.class.getSimpleName();
 
     private static final UnityInterstitialCallbackRouter interstitialRouter = new UnityInterstitialCallbackRouter();
-    private static final UnityBannerCallbackRouter bannerRouter = new UnityBannerCallbackRouter();
 
     static boolean initUnityAds(Map<String, String> serverExtras, Activity launcherActivity) {
         initGdpr(launcherActivity);
@@ -35,9 +33,7 @@ public class UnityRouter {
             return false;
         }
         initMediationMetadata(launcherActivity);
-        UnityBanners.setBannerListener(bannerRouter);
-
-
+        
         boolean testMode = false;
         boolean enablePerPlacementLoad = true;
         UnityAds.initialize(launcherActivity, gameId, interstitialRouter, testMode, enablePerPlacementLoad);
@@ -89,10 +85,6 @@ public class UnityRouter {
 
     static UnityInterstitialCallbackRouter getInterstitialRouter() {
         return interstitialRouter;
-    }
-
-    static UnityBannerCallbackRouter getBannerRouter() {
-        return bannerRouter;
     }
 
     static final class UnityAdsUtils {